requirements.txt
pip==21.2.4
wheel==0.37.0
rsa==4.7.2
pyasn1==0.4.8
six==1.15.0
gast==0.4.0
numpy==1.19.5
h5py==3.1.0
pytz==2021.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
dataclasses==0.6
Jinja2==3.0.2
click==8.0.3
itsdangerous==2.0.1
tensorflow==2.5.0
wrapt==1.12.1
requests==2.26.0
setuptools==58.5.3
geomet==0.2.1.post1
cachetools==4.2.4
certifi==2021.10.8
urllib3==1.26.7
pandas==1.2.0
python-dateutil==2.8.2
SQLAlchemy==1.4.26
greenlet==1.1.2
Markdown==3.3.4
oauthlib==3.1.1
pandasql==0.7.3
cassandra-driver==3.25.0
Flask==2.0.1
tensorboard==2.7.0
flatbuffers==1.12
astunparse==1.6.3
termcolor==1.1.0
zipp==3.6.0
tabulate==0.8.9
Flask-CORS==3.0.10
waitress==2.0.0
duckdb==0.8.1
PyYAML==6.0
ecs-logging==1.1.0
openpyxl==3.0.9
PuLP==2.6.0
protobuf==3.20.0
beautifulsoup4==4.11.1
environs==9.5.0
PyJWT==2.4.0
idna==3.3
cryptography==38.0.1
croniter==1.3.8
python-crontab==2.7.1
marshmallow==3.19.0
azure-storage-blob==1.5.0
azure-storage-common~=1.4
pysftp==0.2.9
XlsxWriter==3.1.2
scipy==1.5.4
scikit-learn==1.3.0
flasgger~=0.9.7.1
Flask-RESTful~=0.3.10
aiohttp~=3.9.1
